requirements.txt
aiohttp~=3.9.1
asyncpg>=0.27.0
click~=8.1.7
discord.py[speed]>=2.2.2
expiringdict~=1.2.2
parsedatetime
Pillow~=10.1.0
psutil~=5.9.6
pygit2~=1.13.3
python-dateutil~=2.8.2
spotipy~=2.23.0
typing_extensions>=4,<5

git+https://github.com/Gorialis/jishaku.git
git+https://github.com/Rapptz/discord-ext-menus.git
wavelink~=3.1.0
fuzzywuzzy~=0.18.0
yarl~=1.9.2
Levenshtein~=0.23.0